server/requirements.txt
fastapi[standard]>=0.110.0
uvicorn>=0.27.1
python-dotenv>=1.0.1
pydantic>=2.6.3
python-multipart>=0.0.9
insightface>=0.7.3
onnxruntime>=1.17.0
opencv-python>=4.9.0
numpy>=1.26.0
fal-client>=0.5.0
httpx>=0.28.0
